+++ templates/expo-template-tabs/app/(tabs)/_layout.tsx
@@ -1,7 +1,8 @@
+import FontAwesome from '@expo/vector-icons/FontAwesome';
 import { Link, Tabs } from 'expo-router';
 import { Pressable, useColorScheme } from 'react-native';
-import Colors from '../../../constants/Colors';
-import FontAwesome from '@expo/vector-icons/FontAwesome';
+
+import Colors from '../../constants/Colors';
 
 /**
  * You can explore the built-in icon families and icons on the web at https://icons.expo.fyi/
